django-master/app/requirements.txt
asgiref==3.3.0
Django==3.0
django-cors-headers==3.5.0
djangorestframework==3.12.1
importlib-metadata==2.0.0
Markdown==3.3.2
pytz==2020.1
sqlparse==0.4.1
zipp==3.3.1